requirements.txt
fastapi
uvicorn
joblib
scikit-learn
numpy
pandas
sqlalchemy
psycopg2-binary
python-jose
bcrypt